requirements.txt
# Core framework
streamlit==1.30.0
python-dotenv==1.0.1

# LangChain ecosystem
langchain==0.3.8
langchain-core==0.3.21
langchain-community==0.3.8
langchain-text-splitters==0.3.2
langchain-google-genai==2.0.4

# Google Gemini (Generative AI)
google-generativeai==0.8.0
protobuf==5.26.1

# Vector store
faiss-cpu==1.8.0

# Document loaders / file parsing
unstructured==0.15.12
PyPDF2==3.0.1
python-docx==1.1.2
requests==2.32.3

# Data handling
pandas==2.2.3
numpy==1.26.4

# Misc utils
tqdm==4.66.5
